requirements.txt
# Database and API connections
supabase>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Web API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# File processing and text extraction
markdown>=3.4.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Web scraping and HTTP requests
requests>=2.31.0
urllib3>=2.0.0

# AI integrations
openai>=1.0.0
google-generativeai>=0.3.0

# Optional: Additional utilities
pathlib2>=2.3.0
lxml>=4.9.0